cmd: add config check for almost all commands
The check has been added for:

- binaries
- build
- check
- clean
- create
- install
- instances
- logrotate
- pack
- restart
- run
- start
- status
- stop
- uninstall

Closes #371

+++ cli/cmd/uninstall.go
@@ -40,6 +40,10 @@ func NewUninstallCmd() *cobra.Command {
 
 // InternalUninstallModule is a default uninstall module.
 func InternalUninstallModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
+	if !isConfigExist(cmdCtx) {
+		return errNoConfig
+	}
+
 	err := uninstall.UninstallProgram(args[0], cliOpts.App.BinDir,
 		cliOpts.App.IncludeDir+"/include", cmdCtx)
 	return err
